+++ src/MapLayer.js
@@ -1,16 +1,16 @@
 import assign from 'lodash/object/assign';
 import React from 'react';
-import MapComponent from './MapComponent'
+import MapComponent from './MapComponent';
 
 export default class MapLayer extends MapComponent {
   componentDidMount() {
     super.componentDidMount();
-    this.props.map.addLayer(this.leafletElement);
+    (this.props.layerGroup || this.props.map).addLayer(this.leafletElement);
   }
 
   componentWillUnmount() {
     super.componentWillUnmount();
-    this.props.map.removeLayer(this.leafletElement);
+    (this.props.layerGroup || this.props.map).removeLayer(this.leafletElement);
   }
 
   getClonedChildrenWithMap(extra) {